Updates to jpeg decoder helper util class
jpegdecoderutil class handles subsampling format 420. Decoding images
with other sub sampling formats return with error. This class is now
updated to handle various 4:a:b configurations

removed possible oob access in parse xmp metadata

fixed warnings in sample app

Test: ./ultrahdr_unit_test
Test: ./ultrahdr_dec_fuzzer

+++ lib/include/ultrahdr/jpegrutils.h
@@ -76,7 +76,7 @@ status_t Write(jr_compressed_ptr destination, const void* source, int length, in
  * @param metadata place to store HDR metadata values
  * @return true if metadata is successfully retrieved, false otherwise
  */
-bool getMetadataFromXMP(uint8_t* xmp_data, size_t xmp_size, ultrahdr_metadata_struct* metadata);
+bool getMetadataFromXMP(uint8_t* xmp_data, int xmp_size, ultrahdr_metadata_struct* metadata);
 
 /*
  * This method generates XMP metadata for the primary image.
